backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pinecone-client
openai
cohere
sentence-transformers
numpy
rank_bm25
pypdf2
tqdm
python-multipart
python-docx
reportlab
ics
supabase
pytesseract
